Skip needless allocations in _TileRenderWrapper
When performing the compositing tile fetch, first check
if the tile with the given coordinate exists in any of
the visible layers of the wrapper. If there is no tile
for the coordinate in any of the layers, return the
constant transparent rgba tile.

Rationale: Although incurring a slight performance hit when
the check fails, a successful check can significantly improve
performance for floodfilling operations when using sample
merged with bg disabled, or filling w. reference to a group.