Optimize largely empty gap closing fills
Enable full-tile skipping when possible; use edge values
in a tuple to represent a full edge of infinite distances,
and permit skipping the fill for a tile with no detected
gaps, which is either the starting tile or reached via
an edge constant.

Enable full-tile replacements even when skips not possible.
When fill is skipped, replace tile with constant full opaque
tile, but also perform replacement when all pixels are filled
at once, regardless (this can happen for e.g corner-crossing gaps)

Update unseeping to account for existence of full-tile constants
and omit full-tile checks at the end of the operation.