Remove tiles when erasing w. fully filled tile
Instead of clearing the tiles, just remove them from
the tile dictionary. This makes erasing fills perform
the function of vacuuming (some) empty tiles on the layer
that it erases from.

Note that this only applies to tiles that were known to
be erased completely - visually empty tiles will not be
removed if the erasing tile was not filled completely.