Add drag-fill functionality to the fill tool
Add the ability to drag-select a range of starting
seeds that the fill is run on, instead of a single
initial seed (still possible; a single click works
the same as before).

The target color is based on the starting position
of the drag.

New overlay indicates the seed pixels when dragging.

Update call chains and the fill functions to handle
seed partitioning and sizing initial bounding boxes.

Update test code.

Add new cursor properties to work with DragMode.
Factor out cursor update from general ui update.