Add pigment default-setting-value switch
Add ability to override setting defaults for BrushInfo objects.
Also track which settings are not defined explicitly on load
(these are guaranteed to start out with default values).

When switching the default on or off, loaded/cached brushes that
did not specify or modify the pigment setting will switch to the
relevant default. Brushes that have the setting set explicitly or
that have been modified are not affected.

The switch callback is handled in the app brushmanager instance,
and the state of the switch is stored in the global preferences.