Make it possible to switch eotf at runtime
Create a new module in lib to handle the EOTF values,
and remove the dependency on get_app.

Replace the stored constant with a function and
initiate the base values from settings on startup.

This setup is obviously not great (global variable).
A sensible refactoring would pass the eotf as a parameter
and have the document models maintain their own eotf.
However, due to the number of chains that would need
updating, this is a necessary (and hopefully temporary)
evil to get working backwards compatibility in time for 2.0.