Make default layer configurable
When using pigment brushes, the natural default layer is
the Pigment layer, but when editing old files the Normal
(src-over) layer would be the sensible default.

Just like the eotf value, this is now a module-global variable
(instead of a module constant) hidden behind a function.
Not great, and should be changed so that it's managed by a doc.

To reduce confusion, the layer list order will remain the same
regardless of the default - the difference is primarily the mode
of new layers.