Add adjustable defaults for unset brush settings
Loaded brushes now keep track of which settings were specified
by the brush, and which are using (new) defaults. Changing a
setting after the brush has been loaded will remove it from the
set of unspecified settings.

This makes it possible to adjust the default values
of brush settings which are not specified in the brush
file at runtime. It is specifically introduced to enable
backwards compability by setting the pigment setting to
off by default when using an eotf.

Keeping track of which settings are unspecified adds a bit
of overhead, but brush loading is usually not done frequently
enough for that to be noticeable.